requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
Faker==12.0.1
frozenlist==1.3.3
idna==3.4
inflection==0.3.1
iniconfig==2.0.0
jsonschema==4.17.3
lazy-object-proxy==1.9.0
mixer==7.2.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
PyJWT==2.6.0
pyrsistent==0.19.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
wrapt==1.15.0
yarl==1.9.2
